requirements.txt
distlib==0.3.8
dnspython==2.7.0
filelock==3.16.1
platformdirs==4.3.6
pymongo==4.8.0
python-dateutil==2.9.0.post0
six==1.16.0
